Lesson 28 - Finish
Animations added for pop/push stack. In typical android fasion, the animations in the text tutorial aren't the dsame as the ones in the video (neither work nicely in a PopExit/Enter/etc).

Checked the git-diff on their real sollution to find a *third* setup of Enter/Exit/PopEnter/PopExit setups....and this one also doesn't seem nice.

Oh well!